requirements.txt
Pillow==10.2.0
spotipy==2.23.0